Add files via upload
PROJECT RECAP:
Social Buzz is a fast growing technology unicorn that need to adapt quickly to it‘s global scale. Accenture has begun a 3 month POC focusing on these tasks: 
An audit of Social Buzz‘s big data practice
Recommendations for a successful IPO
Analysis to find Social Buzz‘s top 5 most popular categories of content 

PROBLEM:
Over 100000 posts per day
36,500,000 pieces of content per year!
But how to capitalize on it when there is so much? 
Analysis to find Social Buzz‘s top 5 most popular categories of content. 

PROCESS:
*Data understanding
*Data cleaning
*Data modeling
*Data Analysis
*Uncover insights

ANALYSIS

Animals and science are the two most popular 
Categories of consent, showing that people enjoy ‘‘real-life‘‘ and ‘‘factual‘‘ content the most. 


INSIGHT

Food is a common theme with the top 5 categories with ‘‘Healthy Eating ‘‘ ranking the highest. This may give an indication to the audience within your user base. You could use this insight to create a campaign and work with healthy eating brands to boost user engagement. 


NEXT STEPS

This ad-hoc analysis is insightful, but it‘s time to take this analysis into large scale production for real-time understanding of your business. We can show you how to do this.